+++ bot/src/main/kotlin/org/epilink/bot/config/LinkConfig.kt
@@ -43,6 +43,10 @@ data class LinkConfiguration(
      */
     val tokens: LinkTokens,
     /**
+     * The OIDC Identity Provider to use
+     */
+    val idProvider: LinkIdProviderConfiguration,
+    /**
      * Discord-related configurations, defined using a [LinkDiscordConfig] object
      */
     val discord: LinkDiscordConfig,
@@ -118,15 +122,25 @@ data class LinkWebServerConfiguration(
      */
     val footers: List<LinkFooterUrl> = listOf(),
     /**
-     * Logo URL that is passed to the front-end
+     * Logo resource asset that is passed to the front-end
      */
-    val logo: String? = null,
+    val logo: ResourceAssetConfig? = ResourceAssetConfig(),
+    /**
+     * Background resource asset that is passed to the front-end
+     */
+    val background: ResourceAssetConfig? = ResourceAssetConfig(),
     /**
      * Contact information for instance maintainers
      *
      * @since 0.2.0
      */
     val contacts: List<LinkContactInformation> = listOf()
+)
+
+data class ResourceAssetConfig(
+    val url: String? = null,
+    val file: String? = null,
+    val contentType: String? = null
 )
 
 /**
@@ -178,17 +192,13 @@ data class LinkTokens(
      */
     val discordOAuthSecret: String,
     /**
-     * Microsoft/Azure AD Client Id
+     * Identity provider OpenID Connect / OAuth 2 Client Id
      */
-    val msftOAuthClientId: String,
+    val idpOAuthClientId: String,
     /**
-     * Microsoft/Azure Ad Client Secret
+     * Identity provider OpenID Connect / OAuth 2 Client Secret
      */
-    val msftOAuthSecret: String,
-    /**
-     * Microsoft tenant. Check the maintainer guide for more information.
-     */
-    val msftTenant: String
+    val idpOAuthSecret: String
 )
 
 /**
@@ -341,6 +351,13 @@ data class LinkLegalConfiguration(
     val identityPromptText: String? = null
 )
 
+data class LinkIdProviderConfiguration(
+    val url: String,
+    val name: String,
+    val icon: ResourceAssetConfig?,
+    val microsoftBackwardsCompatibility: Boolean = false
+)
+
 private val yamlKotlinMapper = ObjectMapper(YAMLFactory()).apply {
     registerModule(KotlinModule())
 }
@@ -372,11 +389,6 @@ fun LinkConfiguration.isConfigurationSane(
     report += discord.check()
     report += discord.checkCoherenceWithRulebook(rulebook)
     report += discord.checkCoherenceWithLanguages(availableDiscordLanguages)
-
-    // Warn if the tenant is broad (common, consumers, organizations) and no e-mail validation is in place
-    if (tokens.msftTenant in setOf("common", "consumers", "organizations") && rulebook.validator == null) {
-        report += ConfigWarning("You are using a non-specific Microsoft tenant (that allows anyone to log in) without e-mail validation: people from domains you do not trust may be accepted by EpiLink!")
-    }
 
     return report
 }
@@ -435,11 +447,11 @@ fun LinkTokens.check(): List<ConfigReportElement> {
     if (discordToken == "...") {
         report += ConfigError(true, "discordToken was left with its default value: please provide a bot token!")
     }
-    if (msftOAuthClientId == "...") {
-        report += ConfigError(true, "msftOAuthClientId was left with its default value: please provide a client ID!")
+    if (idpOAuthClientId == "...") {
+        report += ConfigError(true, "idpOAuthClientId was left with its default value: please provide a client ID!")
     }
-    if (msftOAuthSecret == "...") {
-        report += ConfigError(true, "msftOAuthSecret was left with its default value: please provide a secret!")
+    if (idpOAuthSecret == "...") {
+        report += ConfigError(true, "idpOAuthSecret was left with its default value: please provide a secret!")
     }
     return report
 }