+++ bot/src/main/kotlin/org/epilink/bot/db/LinkGdprReport.kt
@@ -9,6 +9,7 @@
 package org.epilink.bot.db
 
 import org.epilink.bot.LinkServerEnvironment
+import org.epilink.bot.config.LinkIdProviderConfiguration
 import org.epilink.bot.config.LinkPrivacy
 import org.koin.core.KoinComponent
 import org.koin.core.inject
@@ -24,7 +25,7 @@ interface LinkGdprReport {
     suspend fun getFullReport(user: LinkUser, requester: String): String
 
     /**
-     * Generate a report on the user's information (Discord ID, Microsoft ID hash, creation timestamp)
+     * Generate a report on the user's information (Discord ID, Identity Provider ID hash, creation timestamp)
      */
     fun getUserInfoReport(user: LinkUser): String
 
@@ -54,6 +55,7 @@ internal class LinkGdprReportImpl : LinkGdprReport, KoinComponent {
     private val banLogic: LinkBanLogic by inject()
     private val privacy: LinkPrivacy by inject()
     private val env: LinkServerEnvironment by inject()
+    private val idpCfg: LinkIdProviderConfiguration by inject()
 
     override suspend fun getFullReport(user: LinkUser, requester: String): String =
         """
@@ -83,7 +85,7 @@ internal class LinkGdprReportImpl : LinkGdprReport, KoinComponent {
         General information stored about your account.
         
         - Discord ID: $discordId
-        - Microsoft ID hash (Base64 URL-safe encoded): ${user.msftIdHash.encodeBase64Url()}
+        - Identity Provider (${idpCfg.name}) ID hash (Base64 URL-safe encoded): ${user.idpIdHash.encodeBase64Url()}
         - Account creation timestamp: $creationDate
         """.trimIndent()
     }
@@ -128,7 +130,7 @@ internal class LinkGdprReportImpl : LinkGdprReport, KoinComponent {
         """.trimMargin()
 
     private suspend fun makeBansList(user: LinkUser): String? {
-        val bans = dbf.getBansFor(user.msftIdHash)
+        val bans = dbf.getBansFor(user.idpIdHash)
         return if (bans.isEmpty()) {
             null
         } else {
@@ -166,5 +168,7 @@ internal class LinkGdprReportImpl : LinkGdprReport, KoinComponent {
             }
         }
     }
+
+    // TODO add report on language preferences
 
 }