+++ bot/src/main/kotlin/org/epilink/bot/http/endpoints/LinkRegistrationApi.kt
@@ -51,7 +51,7 @@ interface LinkRegistrationApi {
 internal class LinkRegistrationApiImpl : LinkRegistrationApi, KoinComponent {
     private val logger = LoggerFactory.getLogger("epilink.api.registration")
     private val discordBackEnd: LinkDiscordBackEnd by inject()
-    private val microsoftBackEnd: LinkMicrosoftBackEnd by inject()
+    private val idProvider: LinkIdentityProvider by inject()
     private val roleManager: LinkRoleManager by inject()
     private val userApi: LinkUserApi by inject()
     private val userCreator: LinkUserCreator by inject()
@@ -90,16 +90,15 @@ internal class LinkRegistrationApiImpl : LinkRegistrationApi, KoinComponent {
                                 "Missing session header",
                                 "reg.msh"
                             )
-
                         )
-                    } else if (discordId == null || discordUsername == null || email == null || microsoftUid == null) {
+                    } else if (discordId == null || discordUsername == null || email == null || idpId == null) {
                         logger.debug {
                             """
                             Incomplete registration process for session ${call.request.header("RegistrationSessionId")}
                             discordId = $discordId
                             discordUsername = $discordUsername
                             email = $email
-                            microsoftUid = $microsoftUid
+                            idpId = $idpId
                             """.trimIndent()
                         }
                         call.respond(
@@ -111,7 +110,7 @@ internal class LinkRegistrationApiImpl : LinkRegistrationApi, KoinComponent {
                         logger.debug { "Completing registration session for $regSessionId" }
                         val options: AdditionalRegistrationOptions =
                             call.receiveCatching() ?: return@post
-                        val u = userCreator.createUser(discordId, microsoftUid, email, options.keepIdentity)
+                        val u = userCreator.createUser(discordId, idpId, email, options.keepIdentity)
                         roleManager.invalidateAllRoles(u.discordId, true)
                         with(userApi) { loginAs(call, u, discordUsername, discordAvatarUrl) }
                         logger.debug { "Completed registration session. $regSessionId logged in and reg session cleared." }
@@ -124,12 +123,12 @@ internal class LinkRegistrationApiImpl : LinkRegistrationApi, KoinComponent {
             }
 
             @ApiEndpoint("POST /register/authcode/discord")
-            @ApiEndpoint("POST /register/authcode/msft")
+            @ApiEndpoint("POST /register/authcode/idProvider")
             post("authcode/{service}") {
                 when (val service = call.parameters["service"]) {
                     null -> error("Invalid service") // Should not happen
                     "discord" -> processDiscordAuthCode(call, call.receive())
-                    "msft" -> processMicrosoftAuthCode(call, call.receive())
+                    "idProvider" -> processIdProviderAuthCode(call, call.receive())
                     else -> {
                         logger.debug { "Attempted to register under unknown service $service" }
                         call.respond(
@@ -148,28 +147,26 @@ internal class LinkRegistrationApiImpl : LinkRegistrationApi, KoinComponent {
     }
 
     /**
-     * Take a Microsoft authorization code, consume it and apply the information retrieve form it to the current
+     * Take a Identity Provider authorization code, consume it and apply the information retrieve form it to the current
      * registration session
      */
-    private suspend fun processMicrosoftAuthCode(call: ApplicationCall, authcode: RegistrationAuthCode) {
+    private suspend fun processIdProviderAuthCode(call: ApplicationCall, authcode: RegistrationAuthCode) {
         val session = call.sessions.getOrSet { RegisterSession() }
-        logger.debug { "Get Microsoft token from authcode ${authcode.code}" }
-        // Get token
-        val token = microsoftBackEnd.getMicrosoftToken(authcode.code, authcode.redirectUri)
+        logger.debug { "Get IDP identity token from authcode ${authcode.code}" }
         // Get information
-        val (id, email) = microsoftBackEnd.getMicrosoftInfo(token)
-        logger.debug { "Processing Microsoft info for registration for $id ($email)" }
-        val adv = perms.isMicrosoftUserAllowedToCreateAccount(id, email)
+        val (id, email) = idProvider.getUserIdentityInfo(authcode.code, authcode.redirectUri)
+        logger.debug { "Processing IDP info for registration for $id ($email)" }
+        val adv = perms.isIdentityProviderUserAllowedToCreateAccount(id, email)
         if (adv is Disallowed) {
-            logger.debug { "Microsoft user $id ($email) is not allowed to create an account: " + adv.reason }
+            logger.debug { "IDP user $id ($email) is not allowed to create an account: " + adv.reason }
             call.respond(
                 HttpStatusCode.BadRequest,
                 AccountCreationNotAllowed.toResponse(adv.reason, adv.reasonI18n, adv.reasonI18nData)
             )
             return
         }
-        logger.debug { "Microsoft registration information OK: continue " }
-        val newSession = session.copy(email = email, microsoftUid = id)
+        logger.debug { "IDP registration information OK: continue " }
+        val newSession = session.copy(email = email, idpId = id)
         call.sessions.set(newSession)
         call.respond(
             ApiSuccessResponse.of(RegistrationContinuation("continue", newSession.toRegistrationInformation()))