+++ bot/src/test/kotlin/org/epilink/bot/AdminTest.kt
@@ -39,7 +39,7 @@ import kotlin.test.*
 
 private class BanImpl(
     override val banId: Int,
-    override val msftIdHash: ByteArray,
+    override val idpIdHash: ByteArray,
     override val expiresOn: Instant?,
     override val issued: Instant,
     override val revoked: Boolean,
@@ -207,7 +207,7 @@ class AdminTest : KoinBaseTest(
         declare(named("admins")) { listOf("adminid") }
         val targetMock = mockk<LinkUser> {
             every { discordId } returns "targetid"
-            every { msftIdHash } returns byteArrayOf(1, 2, 3)
+            every { idpIdHash } returns byteArrayOf(1, 2, 3)
             every { creationDate } returns instant
         }
         mockHere<LinkDatabaseFacade> {
@@ -226,7 +226,7 @@ class AdminTest : KoinBaseTest(
                 info.data.apply {
                     assertEquals("targetid", get("discordId"))
                     val expectedHash = Base64.getUrlEncoder().encodeToString(byteArrayOf(1, 2, 3))
-                    assertEquals(expectedHash, get("msftIdHash"))
+                    assertEquals(expectedHash, get("idpIdHash"))
                     assertEquals(instant.toString(), get("created"))
                     assertEquals(true, get("identifiable"))
                 }